refactor: extract event handler to view model layer
